+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - GP2-2279 - The progress bar is not displayed on the export plan landing page
 - GP2-2137 - VFM Companies house lookup
 - GP2-2298 - Added header missing menu items (search, services)
+- GP2-2308 - Add new CMS page type that can be used for Services landing page
 - GP2-2155 - Add/Port /healthcheck/ views
 - GP2-2156 - Add robots.txt
 - GP2-2069 - Replicate the Cookie management behaviour from V1.